Criação de domain service
