+++ backend/src/domain/entity/Ride.ts
@@ -1,6 +1,7 @@
 import Position from "./Position";
 import Coord from "../vo/Coord";
 import UUID from "../vo/UUID";
+import DistanceCalculator from "../service/DistanceCalculator";
 
 export default class Ride {
     private rideId: UUID;
@@ -59,7 +60,7 @@ export default class Ride {
 		for (const [index, position] of this.positions.entries()) {
 			const nextPosition = this.positions[index + 1];
 			if (!nextPosition) break;
-			distance += this.calculateDistance(position.getCoord(), nextPosition.getCoord());
+			distance += DistanceCalculator.calculate(position.getCoord(), nextPosition.getCoord());
 		}
         return distance
     }
@@ -86,21 +87,5 @@ export default class Ride {
 
     getDriverId() {
         return this.driverId?.getValue();
-    }
-
-    private calculateDistance(from: Coord, to: Coord) {
-		const earthRadius = 6371;
-		const degreesToRadians = Math.PI / 180;
-		const deltaLat = (to.getLat() - from.getLat()) * degreesToRadians;
-		const deltaLon = (to.getLong() - from.getLong()) * degreesToRadians;
-		const a =
-			Math.sin(deltaLat / 2) * Math.sin(deltaLat / 2) +
-			Math.cos(from.getLat() * degreesToRadians) *
-			Math.cos(to.getLat() * degreesToRadians) *
-			Math.sin(deltaLon / 2) *
-			Math.sin(deltaLon / 2);
-		const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-		const distance = earthRadius * c;
-		return Math.round(distance);
-	}
+    }    
 }
